# public/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "TODO LIST API",
        "contact": {
            "name": "Rafael Pedrosa",
            "url": "https://www.rafadpedrosa.com.br"
        },
        "version": "0.1"
    },
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/authenticate": {
            "post": {
                "tags": [
                    "generic"
                ],
                "summary": "Authentica usu\u00e1rios do sistema",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "usu\u00e1rio authenticado"
                    },
                    "304": {
                        "description": "Usu\u00e1rio e ou senha errads"
                    },
                    "401": {
                        "description": "N\u00e3o autorizado"
                    },
                    "500": {
                        "description": "Erro inesperado"
                    }
                }
            }
        },
        "/api/authenticateByToken": {
            "post": {
                "tags": [
                    "generic"
                ],
                "summary": "Authentica usu\u00e1rios do sistema pelo token",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth_api"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "usu\u00e1rio authenticado"
                    },
                    "304": {
                        "description": "Usu\u00e1rio e ou senha errads"
                    },
                    "401": {
                        "description": "N\u00e3o autorizado"
                    },
                    "500": {
                        "description": "Erro inesperado"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "generic"
                ],
                "summary": "Desloga o usu\u00e1rios do sistema",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth_api"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "usu\u00e1rio deslogado"
                    },
                    "500": {
                        "description": "Erro inesperado"
                    }
                }
            }
        },
        "/api/tasks": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Busca todos os tasks paginados",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/pFilter"
                    },
                    {
                        "$ref": "#/parameters/pPage"
                    },
                    {
                        "$ref": "#/parameters/pPer_page"
                    },
                    {
                        "$ref": "#/parameters/pSort"
                    },
                    {
                        "$ref": "#/parameters/pColumns"
                    },
                    {
                        "$ref": "#/parameters/pPageName"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna tasks paginados"
                    },
                    "500": {
                        "description": "Erro inesperado"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "scope"
                        ]
                    }
                ]
            }
        },
        "/api/task": {
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Salva task",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "task salvo com sucesso"
                    },
                    "500": {
                        "description": "Internal server Error"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "scope"
                        ]
                    }
                ]
            }
        },
        "/api/task/{id}": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Busca task",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/pId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "task encontrado"
                    },
                    "500": {
                        "description": "Internal server Error"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "scope"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "tasks"
                ],
                "summary": "Altera task",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/pId"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "task alterado com sucesso"
                    },
                    "500": {
                        "description": "Internal server Error"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "scope"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "tasks"
                ],
                "summary": "Deleta task",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/pId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "task alterado com sucesso"
                    },
                    "500": {
                        "description": "Internal server Error"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "scope"
                        ]
                    }
                ]
            }
        },
        "/api/task/{id}/edit": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Busca task para ser editado",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/pId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "task encontrado para ser editado"
                    },
                    "500": {
                        "description": "Internal server Error"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "scope"
                        ]
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Busca todos os users paginados",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/pFilter"
                    },
                    {
                        "$ref": "#/parameters/pPage"
                    },
                    {
                        "$ref": "#/parameters/pPer_page"
                    },
                    {
                        "$ref": "#/parameters/pSort"
                    },
                    {
                        "$ref": "#/parameters/pColumns"
                    },
                    {
                        "$ref": "#/parameters/pPageName"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna users paginados"
                    },
                    "500": {
                        "description": "Erro inesperado"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "scope"
                        ]
                    }
                ]
            }
        },
        "/api/user": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Salva user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user salvo com sucesso"
                    },
                    "500": {
                        "description": "Internal server Error"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "scope"
                        ]
                    }
                ]
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Busca user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/pId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user encontrado"
                    },
                    "500": {
                        "description": "Internal server Error"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "scope"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Altera user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/pId"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user alterado com sucesso"
                    },
                    "500": {
                        "description": "Internal server Error"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "scope"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Deleta user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/pId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user alterado com sucesso"
                    },
                    "500": {
                        "description": "Internal server Error"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "scope"
                        ]
                    }
                ]
            }
        },
        "/api/user/{id}/edit": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Busca user para ser editado",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/pId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user encontrado para ser editado"
                    },
                    "500": {
                        "description": "Internal server Error"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "scope"
                        ]
                    }
                ]
            }
        },
        "/api": {
            "get": {
                "summary": "Verificar se a aplica\u00e7\u00e3o est\u00e1 funcionando",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "An print showing laravel lumen message"
                    }
                }
            }
        }
    },
    "definitions": {
        "Task": {
            "properties": {
                "title": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            },
            "type": "object",
            "example": {
                "title": "this is a title",
                "description": "this is a description",
                "userId": "8"
            }
        },
        "User": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            },
            "type": "object",
            "example": {
                "name": "User name",
                "description": "User 1"
            }
        },
        "SWGGeneric": {
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            },
            "type": "object",
            "example": {
                "name": "test"
            }
        },
        "SWGUserLogin": {
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            },
            "type": "object",
            "example": {
                "email": "1@1.com",
                "password": "1"
            }
        },
        "auth": {
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            },
            "type": "object",
            "example": {
                "email": "User email",
                "password": "Senha do usu\u00e1rio"
            }
        },
        "auth_api": {
            "properties": {
                "api_key": {
                    "type": "string"
                }
            },
            "type": "object",
            "example": {
                "api_key": "api string here"
            }
        }
    },
    "parameters": {
        "generic_def": {
            "name": "generic",
            "in": "body",
            "description": "JSON obj",
            "required": true,
            "schema": {
                "$ref": "#/definitions/SWGGeneric"
            }
        },
        "user_login": {
            "name": "user",
            "in": "body",
            "description": "JSON obj",
            "required": true,
            "schema": {
                "$ref": "#/definitions/SWGUserLogin"
            }
        },
        "pAuthorization": {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
        },
        "pEmail": {
            "name": "email",
            "in": "body",
            "required": true,
            "schema": {
                "title": "json",
                "properties": {
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "email": "1@1.com"
                }
            },
            "type": "string"
        },
        "pFilter": {
            "name": "filter",
            "in": "query",
            "description": "query string filter",
            "type": "string"
        },
        "pId": {
            "name": "id",
            "in": "path",
            "description": "path - id",
            "type": "string"
        },
        "pPage": {
            "name": "page",
            "in": "query",
            "description": "query string page",
            "type": "string"
        },
        "pPer_page": {
            "name": "per_page",
            "in": "query",
            "description": "query string per_page",
            "type": "string"
        },
        "pSort": {
            "name": "sort",
            "in": "query",
            "description": "query string sort",
            "type": "string"
        },
        "pColumns": {
            "name": "columns",
            "in": "query",
            "description": "query string columns",
            "type": "string"
        },
        "pPageName": {
            "name": "pageName",
            "in": "query",
            "description": "query string page name",
            "type": "string"
        }
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "scopes": {
                "scope": "Description of scope."
            }
        }
    },
    "tags": [
        {
            "name": "diversos",
            "description": "Opera\u00e7\u00f5es diversas do sistema",
            "externalDocs": {
                "description": "",
                "url": ""
            }
        },
        {
            "name": "config",
            "description": "Toda a api de config",
            "externalDocs": {
                "description": "",
                "url": ""
            }
        },
        {
            "name": "generic",
            "description": "AJUSTE AQUI",
            "externalDocs": {
                "description": "",
                "url": ""
            }
        },
        {
            "name": "users",
            "description": "users  API description",
            "externalDocs": {
                "description": "",
                "url": ""
            }
        },
        {
            "name": "tasks",
            "description": "tasks  API description",
            "externalDocs": {
                "description": "",
                "url": ""
            }
        }
    ]
}